double no stopping criterion

+++ Chart.java
@@ -1,7 +1,6 @@
 import java.awt.Color;
 import java.io.File;
 import java.io.IOException;
-import java.math.BigDecimal;
 import java.util.List;
 
 import org.jfree.chart.ChartFactory;
@@ -12,13 +11,13 @@ import org.jfree.data.xy.XYSeries;
 import org.jfree.data.xy.XYSeriesCollection;
 
 public class Chart {
-   private final List<BigDecimal> meanList;
+   private final List<Double> meanList;
    private final String fileName;
    private XYSeries series;
    private XYSeriesCollection dataset;
    private JFreeChart chart;
 
-   public Chart(List<BigDecimal> meanList, String fileName) {
+   public Chart(List<Double> meanList, String fileName) {
       this.meanList = meanList;
       this.fileName = fileName;
       series = new XYSeries("Data Series");
@@ -26,7 +25,7 @@ public class Chart {
    }
    
    private void draw() {
-      for (BigDecimal mean_value : meanList) {
+      for (double mean_value : meanList) {
          series.add(meanList.indexOf(mean_value), mean_value);
       }
       dataset = new XYSeriesCollection(series);
